Class for continuous anemometer sampling
Allows continuous sampling for a period of time in the loop
Note: Set SMP_LOOP_OFFSET to be 65 ms, meaning we have 35 ms to sample and log (write takes ~6 ms)
contSamp Class combines and re-purposes elements of ADCSampler and Logger.  It has not been compiled yet to check for errors
I am also not 100% sure that we can initilize the SD card twice.  I think it makes sense though that we could be writing to different files at different times.